requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
jinja2>=3.0.0
python-multipart>=0.0.5
pydantic>=1.8.0
pydantic-settings==2.0.3
aiofiles>=0.7.0
aiohttp==3.8.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pycryptodome==3.19.0
apscheduler==3.10.4
sqlalchemy>=1.4.0
alembic>=1.7.0
aiosqlite>=0.17.0
python-dotenv>=0.19.0
email-validator>=1.1.0
websockets>=10.0
pillow==10.1.0
python-socketio>=5.4.0 